refactor: update all imports, tests, examples, and docs for new categories
Updates imports across unit tests, integration tests, examples, benchmarks,
CLI, solver, and source doc comments from old categories (optimization,
satisfiability, specialized) to new ones (algebraic, formula, graph, misc).
