+++ examples/reduction_ilp_to_qubo.rs
@@ -36,7 +36,7 @@
 // ```
 
 use problemreductions::export::*;
-use problemreductions::models::optimization::{LinearConstraint, ObjectiveSense, ILP};
+use problemreductions::models::algebraic::{LinearConstraint, ObjectiveSense, ILP};
 use problemreductions::prelude::*;
 
 pub fn run() {
