+++ examples/reduction_maximumsetpacking_to_ilp.rs
@@ -15,7 +15,7 @@
 // Exports `docs/paper/examples/maximumsetpacking_to_ilp.json` and `maximumsetpacking_to_ilp.result.json`.
 
 use problemreductions::export::*;
-use problemreductions::models::optimization::ILP;
+use problemreductions::models::algebraic::ILP;
 use problemreductions::prelude::*;
 
 pub fn run() {
