+++ examples/reduction_minimumdominatingset_to_ilp.rs
@@ -14,7 +14,7 @@
 // Exports `docs/paper/examples/minimumdominatingset_to_ilp.json` and `minimumdominatingset_to_ilp.result.json`.
 
 use problemreductions::export::*;
-use problemreductions::models::optimization::ILP;
+use problemreductions::models::algebraic::ILP;
 use problemreductions::prelude::*;
 use problemreductions::topology::small_graphs::petersen;
 use problemreductions::topology::{Graph, SimpleGraph};
