+++ src/models/algebraic/ilp.rs
@@ -195,7 +195,7 @@ pub enum ObjectiveSense {
 /// # Example
 ///
 /// ```
-/// use problemreductions::models::optimization::{ILP, VarBounds, Comparison, LinearConstraint, ObjectiveSense};
+/// use problemreductions::models::algebraic::{ILP, VarBounds, Comparison, LinearConstraint, ObjectiveSense};
 /// use problemreductions::Problem;
 ///
 /// // Create a simple ILP: maximize x0 + 2*x1
