+++ src/models/formula/sat.rs
@@ -89,7 +89,7 @@ impl CNFClause {
 /// # Example
 ///
 /// ```
-/// use problemreductions::models::satisfiability::{Satisfiability, CNFClause};
+/// use problemreductions::models::formula::{Satisfiability, CNFClause};
 /// use problemreductions::{Problem, Solver, BruteForce};
 ///
 /// // Formula: (x1 OR x2) AND (NOT x1 OR x3) AND (NOT x2 OR NOT x3)
