+++ src/unit_tests/solvers/ilp/solver.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::models::optimization::{LinearConstraint, VarBounds};
+use crate::models::algebraic::{LinearConstraint, VarBounds};
 use crate::solvers::BruteForce;
 use crate::traits::Problem;
